requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cytoolz==0.12.0
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
hexbytes==0.3.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
lxml==4.9.1
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.23.5
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
parsimonious==0.8.1
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.5
pycryptodome==3.16.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rlp==2.0.1
scipy==1.9.3
six==1.16.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
tweepy==4.12.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
varint==1.0.2
web3==5.31.1
websockets==9.1
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
(venv) connectfast@Koens-MBP code % pip freeze
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cytoolz==0.12.0
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
hexbytes==0.3.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
lxml==4.9.1
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.23.5
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
parsimonious==0.8.1
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.5
pycryptodome==3.16.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rlp==2.0.1
scipy==1.9.3
six==1.16.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
tweepy==4.12.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
varint==1.0.2
web3==5.31.1
websockets==9.1
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0